Update `executedMovesTotal` to use `BigInteger` for better precision and adjust related calculations and formatting

+++ epd-core/src/main/java/net/chesstango/epd/core/report/SummaryModel.java
@@ -13,6 +13,7 @@ import net.chesstango.search.SearchResultByDepth;
 import net.chesstango.search.SearchResult;
 
 
+import java.math.BigInteger;
 import java.util.HashMap;
 import java.util.LinkedList;
 import java.util.List;
@@ -42,7 +43,7 @@ public class SummaryModel {
     int depthAccuracyAvgPercentageTotal;
 
     @JsonProperty("executedMovesTotal")
-    long executedMovesTotal;
+    BigInteger executedMovesTotal;
 
     @JsonProperty("maxSearchRLevel")
     int maxSearchRLevel;
@@ -128,7 +129,7 @@ public class SummaryModel {
         model.visitedRNodesTotal = nodesReportModel.visitedRNodesTotal;
         model.visitedQNodesTotal = nodesReportModel.visitedQNodesTotal;
         model.visitedNodesTotal = nodesReportModel.visitedNodesTotal;
-        model.executedMovesTotal = nodesReportModel.executedMovesTotal;
+        model.executedMovesTotal = BigInteger.valueOf(nodesReportModel.executedMovesTotal);
         model.cutoffPercentageTotal = nodesReportModel.cutoffPercentageTotal;
         model.evaluationCounterTotal = evaluationReportModel.evaluationCounterTotal;
         model.evaluationCollisionPercentageTotal = evaluationReportModel.evaluationCollisionPercentageTotal;
